[#6] fix: 회원가입 코드리뷰반영 - 테스트 코드 패키지 구조 변경
